feat: token test update
