+++ apps/galaxy/src/utils/tokenService.ts
@@ -1,37 +1,32 @@
-export default function tokenService() {
-  const setToken = (token: string) => {
-    console.log("setToken");
-    localStorage.setItem("accessToken", token);
-  };
+export const setToken = (token: string) => {
+  localStorage.setItem("accessToken", token);
+};
 
-  const getToken = () => {
-    if (typeof window !== "undefined") {
-      return localStorage.getItem("accessToken");
-    }
-  };
-
-  const checkTokenExpiration = () => {
-    const token = getToken();
-    const utcTimestamp = new Date();
-    return utcTimestamp > parseJwtToken(token as string);
-  };
-
-  function parseJwtToken(token: string) {
-    const base64Url = token.split(".")[1];
-    const base64 = base64Url.replace(/-/g, "+").replace(/_/g, "/");
-    const jsonPayload = decodeURIComponent(
-      window
-        .atob(base64)
-        .split("")
-        .map(function (c) {
-          return "%" + ("00" + c.charCodeAt(0).toString(16)).slice(-2);
-        })
-        .join("")
-    );
-    const exp = new Date(0);
-    exp.setUTCSeconds(JSON.parse(jsonPayload).exp);
-    return exp;
+export const getToken = () => {
+  if (typeof window !== "undefined") {
+    return localStorage.getItem("accessToken");
   }
+};
 
-  return { setToken, getToken, checkTokenExpiration };
-}
+export const checkTokenExpiration = () => {
+  const token = getToken();
+  const utcTimestamp = new Date();
+  return utcTimestamp > parseJwtToken(token as string);
+};
+
+const parseJwtToken = (token: string) => {
+  const base64Url = token.split(".")[1];
+  const base64 = base64Url.replace(/-/g, "+").replace(/_/g, "/");
+  const jsonPayload = decodeURIComponent(
+    window
+      .atob(base64)
+      .split("")
+      .map(function (c) {
+        return "%" + ("00" + c.charCodeAt(0).toString(16)).slice(-2);
+      })
+      .join("")
+  );
+  const exp = new Date(0);
+  exp.setUTCSeconds(JSON.parse(jsonPayload).exp);
+  return exp;
+};
